requirements.txt
fastapi>=0.120.4
uvicorn[standard]>=0.38.0
python-multipart>=0.0.20
aiofiles>=25.1.0
google-generativeai>=0.3.0
PyPDF2>=3.0.1
python-docx>=1.2.0
sqlalchemy>=2.0.44
pydantic>=2.12.3
chromadb>=1.3.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.5.0
email-validator>=2.3.0
psycopg2-binary>=2.9.11
alembic>=1.17.1
python-dotenv>=1.0.0
